Add login/signup functionality to about.php, logout functionality
The about page was given the login and signup functionality with include statements.  I do not have access to the about page in the shared folder so it will need to be added once I get that access on FileZilla.  Logout functionality was created and can be added to future pages.