Join event functionality
Added functionality to join event so that the query is processed, and will send a pop up about if it was successful or not, then redirecting back to the FindEvents.php page.  Also, on FindEvents.php I switched the order so that name is the first column, and the join button is in the right column.